Members cleanup and review fixes.

+++ src/modelconfig.cpp
@@ -422,10 +422,10 @@ Status ModelConfig::parseNode(const rapidjson::Value& v) {
         this->setLowLatencyTransformation(v["low_latency_transformation"].GetBool());
 
     if (v.HasMember("sequence_timeout"))
-        this->setSequenceTimeout(v["sequence_timeout"].GetUint64());
+        this->setSequenceTimeout(v["sequence_timeout"].GetUint32());
 
     if (v.HasMember("max_sequence_number"))
-        this->setMaxSequenceNumber(v["max_sequence_number"].GetUint64());
+        this->setMaxSequenceNumber(v["max_sequence_number"].GetUint32());
 
     if (v.HasMember("model_version_policy")) {
         rapidjson::StringBuffer buffer;
